Fixed name space issues that were caused by moving files
