+++ Assets/Scripts/Files/Directory.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 
-namespace FileStructure
+namespace Files
 {
     public class Directory : IFileSystemItem
     {
